# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/group/upsert": {
            "post": {
                "operationId": "ef0b3b6cbf58ade9897d3c95dbb6d175",
                "parameters": [
                    {
                        "name": "X-Custom-Header",
                        "in": "header",
                        "description": "Custom header for additional information",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "categoryid",
                                    "groupname"
                                ],
                                "properties": {
                                    "categoryid": {
                                        "description": "Category ID",
                                        "type": "string"
                                    },
                                    "groupid": {
                                        "description": "Group ID (when provided, item will be updated, otherwise created)",
                                        "type": "string"
                                    },
                                    "groupname": {
                                        "description": "Group Name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully Saved"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Invalid Token (Usually Occurs when token is expired, which requires re-login)"
                    },
                    "404": {
                        "description": "Invalid Category ID <br/>Invalid Group ID"
                    }
                }
            }
        },
        "/api/group/delete": {
            "post": {
                "operationId": "cc57f38b19a079e3c71173250e29eba3",
                "parameters": [
                    {
                        "name": "X-Custom-Header",
                        "in": "header",
                        "description": "Custom header for additional information",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "groupid",
                                    "categoryid"
                                ],
                                "properties": {
                                    "groupid": {
                                        "description": "Group ID",
                                        "type": "string"
                                    },
                                    "categoryid": {
                                        "description": "Category ID",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully Saved"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Invalid Token (Usually Occurs when token is expired, which requires re-login)"
                    },
                    "404": {
                        "description": "Invalid Category ID <br/>Invalid Group"
                    }
                }
            }
        },
        "/api/group/get/": {
            "get": {
                "summary": "Get Groups",
                "operationId": "a3f104ec56ae54333fc2bb7c6d62ca2a",
                "parameters": [
                    {
                        "name": "categoryid",
                        "in": "query",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "groupid",
                        "in": "query",
                        "description": "Group ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Custom-Header",
                        "in": "header",
                        "description": "Custom header for additional information",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Body"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/group/getbyclient/": {
            "get": {
                "summary": "Get Groups by client",
                "operationId": "e8ef3ffe9c52dfe4416851d1a798c3a3",
                "parameters": [
                    {
                        "name": "clientcode",
                        "in": "query",
                        "description": "Client Code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "groupid",
                        "in": "query",
                        "description": "Group ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Custom-Header",
                        "in": "header",
                        "description": "Custom header for additional information",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Body"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    }
}